Moves into style folder
