+++ src/components/StyleComponents/NoiseOverlay.vue
